requirements.txt
flask
openai
gunicorn
flask_login
mysql-connector-python